- feat: edit size image - feat: move search suggestion item component to a new file
